+++ packages/EasyApiPlatform/composer.json
@@ -6,15 +6,15 @@
     "require": {
         "php": "^8.1",
         "api-platform/core": "^2.6 || ^2.7",
+        "doctrine/dbal": "^2.6",
+        "doctrine/doctrine-bundle": "^2.6",
+        "doctrine/orm": "^2.6",
+        "monolog/monolog": "^1.22 || ^2.0",
         "symfony/dependency-injection": "^5.4 || ^6.0",
         "symfony/http-foundation": "^5.4 || ^6.0"
     },
     "require-dev": {
-        "doctrine/dbal": "^2.6",
-        "doctrine/doctrine-bundle": "^2.6",
-        "doctrine/orm": "^2.6",
         "mockery/mockery": "^1.4",
-        "monolog/monolog": "^1.22 || ^2.0",
         "phpspec/prophecy": "^1.0",
         "phpspec/prophecy-phpunit": "^2.0",
         "phpunit/phpunit": "^9.5",
