+++ packages/EasyApiPlatform/src/Bridge/Symfony/DependencyInjection/EasyApiPlatformExtension.php
@@ -22,5 +22,6 @@ final class EasyApiPlatformExtension extends Extension
 
         $loader = new PhpFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.php');
+        $loader->load('filters.php');
     }
 }
