+++ packages/EasyApiPlatform/tests/Filters/AbstractFilterTestCase.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Tests\Filter;
+namespace EonX\EasyApiPlatform\Tests\Filters;
 
 use ApiPlatform\Doctrine\Orm\Filter\FilterInterface;
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGenerator;
